+++ hc/api/tests/test_pause.py
@@ -6,31 +6,30 @@ from hc.test import BaseTestCase
 
 
 class PauseTestCase(BaseTestCase):
-    def test_it_works(self):
-        check = Check.objects.create(project=self.project, status="up")
+    def setUp(self):
+        super().setUp()
 
-        url = "/api/v1/checks/%s/pause" % check.code
+        self.check = Check.objects.create(project=self.project, status="up")
+        self.url = f"/api/v1/checks/{self.check.code}/pause"
+
+    def test_it_works(self):
         r = self.client.post(
-            url, "", content_type="application/json", HTTP_X_API_KEY="X" * 32
+            self.url, "", content_type="application/json", HTTP_X_API_KEY="X" * 32
         )
 
         self.assertEqual(r.status_code, 200)
         self.assertEqual(r["Access-Control-Allow-Origin"], "*")
 
-        check.refresh_from_db()
-        self.assertEqual(check.status, "paused")
+        self.check.refresh_from_db()
+        self.assertEqual(self.check.status, "paused")
 
     def test_it_handles_options(self):
-        check = Check.objects.create(project=self.project, status="up")
-
-        r = self.client.options("/api/v1/checks/%s/pause" % check.code)
+        r = self.client.options(self.url)
         self.assertEqual(r.status_code, 204)
         self.assertIn("POST", r["Access-Control-Allow-Methods"])
 
     def test_it_only_allows_post(self):
-        url = "/api/v1/checks/1659718b-21ad-4ed1-8740-43afc6c41524/pause"
-
-        r = self.client.get(url, HTTP_X_API_KEY="X" * 32)
+        r = self.client.get(self.url, HTTP_X_API_KEY="X" * 32)
         self.assertEqual(r.status_code, 405)
 
     def test_it_validates_ownership(self):
@@ -60,19 +59,29 @@ class PauseTestCase(BaseTestCase):
         self.assertEqual(r.status_code, 404)
 
     def test_it_clears_last_start_alert_after(self):
-        check = Check(project=self.project, status="up")
-        check.last_start = now()
-        check.alert_after = check.last_start + td(hours=1)
-        check.save()
+        self.check.last_start = now()
+        self.check.alert_after = self.check.last_start + td(hours=1)
+        self.check.save()
 
-        url = "/api/v1/checks/%s/pause" % check.code
         r = self.client.post(
-            url, "", content_type="application/json", HTTP_X_API_KEY="X" * 32
+            self.url, "", content_type="application/json", HTTP_X_API_KEY="X" * 32
         )
 
         self.assertEqual(r.status_code, 200)
         self.assertEqual(r["Access-Control-Allow-Origin"], "*")
 
-        check.refresh_from_db()
-        self.assertEqual(check.last_start, None)
-        self.assertEqual(check.alert_after, None)
+        self.check.refresh_from_db()
+        self.assertEqual(self.check.last_start, None)
+        self.assertEqual(self.check.alert_after, None)
+
+    def test_it_clears_next_nag_date(self):
+        self.profile.nag_period = td(hours=1)
+        self.profile.next_nag_date = now() + td(minutes=30)
+        self.profile.save()
+
+        self.client.post(
+            self.url, "", content_type="application/json", HTTP_X_API_KEY="X" * 32
+        )
+
+        self.profile.refresh_from_db()
+        self.assertIsNone(self.profile.next_nag_date)