+++ hc/front/views.py
@@ -535,6 +535,10 @@ def pause(request, code):
     check.alert_after = None
     check.save()
 
+    # After pausing a check we must check if all checks are up,
+    # and Profile.next_nag_date needs to be cleared out:
+    check.project.update_next_nag_dates()
+
     # Don't redirect after an AJAX request:
     if request.META.get("HTTP_X_REQUESTED_WITH") == "XMLHttpRequest":
         return HttpResponse()